+++ docs/api.rst
@@ -4,17 +4,17 @@ API Reference
 Pre-processing scripts
 ----------------------
 
-.. automodule:: vasp.pre_processing.automated_preprocessing
+.. automodule:: vasp.preprocessing.preprocessing
    :members:
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: vasp.pre_processing.preprocessing_common
+.. automodule:: vasp.preprocessing.preprocessing_common
    :members:
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: vasp.pre_processing.vmtkmeshgeneratorfsi
+.. automodule:: vasp.preprocessing.vmtkmeshgeneratorfsi
    :members:
    :undoc-members:
    :show-inheritance:
@@ -23,61 +23,61 @@ Pre-processing scripts
 Post-processing scripts
 -----------------------
 
-.. automodule:: vasp.post_processing.log_plotter
+.. automodule:: vasp.postprocessing.log_plotter
    :members:
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: vasp.post_processing.predeform_mesh
+.. automodule:: vasp.postprocessing.predeform_mesh
    :members:
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: vasp.post_processing.postprocessing_common
+.. automodule:: vasp.postprocessing.postprocessing_common
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_fenics.create_hdf5
+.. automodule:: vasp.postprocessing.postprocessing_fenics.create_hdf5
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_fenics.create_separate_domain_visualization
+.. automodule:: vasp.postprocessing.postprocessing_fenics.create_separate_domain_visualization
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_fenics.postprocessing_fenics_common
+.. automodule:: vasp.postprocessing.postprocessing_fenics.postprocessing_fenics_common
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_h5py.chroma_filters
+.. automodule:: vasp.postprocessing.postprocessing_h5py.chroma_filters
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_h5py.create_spectrograms_chromagrams
+.. automodule:: vasp.postprocessing.postprocessing_h5py.create_spectrograms_chromagrams
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_h5py.create_spectrum
+.. automodule:: vasp.postprocessing.postprocessing_h5py.create_spectrum
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_h5py.postprocessing_common_h5py
+.. automodule:: vasp.postprocessing.postprocessing_h5py.postprocessing_h5py_common
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_h5py.spectrograms
+.. automodule:: vasp.postprocessing.postprocessing_h5py.spectrograms
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_mesh.postprocessing_mesh_common
+.. automodule:: vasp.postprocessing.postprocessing_mesh.postprocessing_mesh_common
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_mesh.separate_mesh
+.. automodule:: vasp.postprocessing.postprocessing_mesh.separate_mesh
    :members:
    :undoc-members:
 
-.. automodule:: vasp.post_processing.postprocessing_mesh.create_refined_mesh
+.. automodule:: vasp.postprocessing.postprocessing_mesh.create_refined_mesh
    :members:
    :undoc-members:
