+++ docs/preprocess.md
@@ -2,7 +2,7 @@
 
 Since pre-processing part of `VaSP` is based on [VaMPy](https://github.com/KVSlab/VaMPy), overlapping functionality of the pre-processing will not be covered here. Instead, users are encouraged to refer to the [VaMPy documentation](https://kvslab.github.io/VaMPy/preprocess.html) for some of the basic functionalities used in `VaSP`. In this document, we will focus on the newly implemented parts of the pre-processing that are specific for performing fluid-structure interaction (FSI) simulations.
 
-The main script for running pre-processing is `automated_preprocessing.py` located under `src/vasp/pre_processing` and can be run by using the command `vasp-generate-mesh`. Although there are quite many parameters for `vasp-generate-mesh`, it is not necessary to specify all the arguments. In fact, it is recommended that the user starts from providing minimally required arguments first, and then control the other arguments based on the user's need for meshing. To run the pre-processing, i.e. generating the volume mesh for FSI simulations, it is required to provide the surface mesh in `vtp` or `stl` format. Assume that the user has the surface mesh named `my_mesh.vtp`. Then, pre-processing can be performed as
+The main script for running pre-processing is `preprocessing.py` located under `src/vasp/preprocessing` and can be run by using the command `vasp-generate-mesh`. Although there are quite many parameters for `vasp-generate-mesh`, it is not necessary to specify all the arguments. In fact, it is recommended that the user starts from providing minimally required arguments first, and then control the other arguments based on the user's need for meshing. To run the pre-processing, i.e. generating the volume mesh for FSI simulations, it is required to provide the surface mesh in `vtp` or `stl` format. Assume that the user has the surface mesh named `my_mesh.vtp`. Then, pre-processing can be performed as
 
 ``` console
 vasp-generate-mesh --input-model /[some_path]/my_mesh.vtp
