+++ src/vasp/postprocessing/postprocessing_fenics/create_separate_domain_visualization.py
