+++ src/vasp/postprocessing/postprocessing_h5py/postprocessing_h5py_common.py
@@ -21,7 +21,7 @@ from tqdm import tqdm
 from scipy import signal
 import matplotlib.pyplot as plt
 
-from vasp.post_processing.postprocessing_common import get_domain_ids, output_file_lists, \
+from vasp.postprocessing.postprocessing_common import get_domain_ids, output_file_lists, \
     read_parameters_from_file
