+++ src/vasp/preprocessing/__init__.py
@@ -1,12 +1,12 @@
 # Copyright (c) 2023 Simula Research Laboratory
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from . import automated_preprocessing
+from . import preprocessing
 from . import preprocessing_common
 from . import generate_solid_probe
 
 __all__ = [
-    "pre_processing",
+    "preprocessing",
     "preprocessing_common",
     "generate_solid_probe",
 ]
