+++ src/vasp/preprocessing/generate_solid_probe.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import numpy as np
 import json
 
-from vasp.post_processing.postprocessing_common import get_domain_ids
+from vasp.postprocessing.postprocessing_common import get_domain_ids
 
 
 def parse_arguments() -> argparse.Namespace:
